resources/scripts/requirements.txt
torch==2.0.0
torchvision==0.15.1
opencv_contrib_python==4.7.0.68
opencv_contrib_python_headless==4.8.1.78
opencv_python==4.8.0.76
opencv_python_headless==4.8.1.78
pycocotools==2.0.6
accelerate
diffusers==0.23.0
matplotlib==3.5.2
numpy==1.24.3
Pillow==10.4.0
rembg==2.0.58
Requests==2.32.3
transformers==4.44.0
notebook==7.0.6
timm==0.4.12
coremltools==7.1
yacs==0.1.8
loralib==0.1.2
kornia==0.7.1
wandb==0.16.3
mmengine==0.10.3
mmcv==2.0.0rc4
mmdet==3.3.0
tensorboard==2.14.0